DUI: fix button custom image icon not being loaded on encrypted device
if the device is fully encrypted, resources in the data partition
can't be loaded till the user fully unlocks the device after boot
inserting the pin, so Fling or SmartBar button icons can't be set
correctly if they are linked to a custom image in that partition.
Navbar code, instead, gets loaded before the unlock steps are done.

This has been noticed on Pixel2xl, shamu and other devices, where for
example the user gets the default DU logo instead of his custom one.

To fix this, we just trigger an icon refresh after the user completes
the unlock process and all files in the data partition are available.

PS: In the future it'd be better to take care of this boot completed event
for all features where we must access personal files on encrypted devices.

Change-Id: Iec5508dba5538ea6530721f40724be1a59db3f22